Projeler/_arsiv/B2B_Outreach/requirements.txt
requests
pydantic
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openai
apify-client
